Add comparison instructions:
This could be done a bit nicer, but because there are so many comparison
instructions there wasn't an easy/more expressive way to express this. At some
point I want to add operator overloads to automatically create these comparison
nodes, though that will add a bit of overhead and complexity.