Add ability to directly run QBE IR:
This allows for users to:

* Convert QBE IR to assembly
* Convert assembly to a binary
* Run final binary

This process will need to be improved (and tested) in the future, but for now
you can use qbew to create runnable programs. One notable feature that is
missing is the ability to create an object file from the IR. This is easy
enough to do if you have the assembly, but there should be a helper function to
do this for you.